Move filter, page and sort code from GET to their helper classes
