+++ src/Controller/GET.php
@@ -17,6 +17,7 @@
 namespace Phramework\JSONAPI\Controller;
 
 use Phramework\JSONAPI\Controller\GET\Filter;
+use Phramework\JSONAPI\Controller\GET\Page;
 use Phramework\JSONAPI\Controller\GET\Sort;
 use \Phramework\Models\Request;
 use \Phramework\Models\Operator;
@@ -35,9 +36,9 @@ abstract class GET extends \Phramework\JSONAPI\Controller\GETById
      * @param  object $params  Request parameters
      * @param  string $modelClass                      Resource's primary model
      * to be used
-     * @param  array $additionalGetArguments           [Optional] Array with any
+     * @param  array $primaryDataParameters           [Optional] Array with any
      * additional arguments that the primary data is requiring
-     * @param  array $additionalRelationshipsArguments [Optional] Array with any
+     * @param  array $relationshipsParameters [Optional] Array with any
      * additional argument primary data's relationships are requiring
      * @param  boolean $filterable                     [Optional] Deafult is
      * true, if true allows `filter` URI parameters to be parsed for filtering
@@ -49,299 +50,44 @@ abstract class GET extends \Phramework\JSONAPI\Controller\GETById
      * @uses $modelClass::get method to fetch resource collection
      * @throws \Exception
      * @throws RequestException
+     * @return boolean
      */
     protected static function handleGET(
         $params,
         $modelClass,
-        $additionalGetArguments = [],
-        $additionalRelationshipsArguments = [],
+        $primaryDataParameters = [],
+        $relationshipsParameters = [],
         $filterable = true,
         $filterableJSON = false,
         $sortable = true
     ) {
-        $page = null;
-
-        $filter = new Filter();
-
-        /*= (object)[
-            'primary' => null,
-            'relationships' => [],
-            'attributes' => [],
-            'attributesJSON' => []
-        ];*/
+        $filter = null;
 
         $sort = null;
 
-        $idAttribute = $modelClass::getIdAttribute();
+        if ($filterable){
 
-        $filterValidationModel = $modelClass::getFilterValidationModel();
-
-        if ($filterable && isset($params->filter)) {
-            foreach ($params->filter as $filterKey => $filterValue) {
-                //todo validate as int
-
-                if ($filterKey === $modelClass::getType()) {
-                    //Check filter value type
-                    if (!is_string($filterValue) && !is_numeric($filterValue)) {
-                        throw new RequestException(sprintf(
-                            'String or integer value required for filter "%s"',
-                            $filterKey
-                        ));
-                    }
-
-                    //Use filterValidator for idAttribute if set else use intval to parse filtered values
-                    $function = (
-                    !empty($filterValidationModel) && isset($filterValidationModel->{$idAttribute})
-                        ? [$filterValidationModel->{$idAttribute}, 'parse']
-                        : 'intval'
-                    );
-
-                    $values = array_map(
-                        $function,
-                        array_map('trim', explode(',', trim($filterValue)))
-                    );
-
-                    $filter->primary = $values;
-                } elseif ($modelClass::relationshipExists($filterKey)) {
-                    //Check filter value type
-                    if (!is_string($filterValue) && !is_numeric($filterValue)) {
-                        throw new RequestException(sprintf(
-                            'String or integer value required for filter "%s"',
-                            $filterKey
-                        ));
-                    }
-
-                    //Todo use filterValidation model
-
-                    $values = array_map(
-                        'intval',
-                        array_map('trim', explode(',', trim($filterValue)))
-                    );
-
-                    $filter->relationships[$filterKey] = $values;
-
-                    //when TYPE_TO_ONE it's easy to filter
-                } else {
-                    $validationModel = $modelClass::getValidationModel();
-                    //if (!$validationModel || !isset($validationModel->attributes)) {
-                    //    throw new \Exception(sprintf(
-                    //        'Model "%s" doesn\'t have a validation model for attributes',
-                    //        $modelClass::getType()
-                    //    ));
-                    //}
-                    $attributeValidationModel = null;
-
-                    if ($filterValidationModel
-                        && isset($filterValidationModel)
-                        && isset($filterValidationModel->properties->{$filterKey})
-                    ) {
-                        $attributeValidationModel =
-                            $filterValidationModel->properties->{$filterKey};
-                    } elseif ($validationModel
-                        && isset($validationModel->attributes)
-                        && isset($validationModel->attributes->properties->{$filterKey})
-                    ) {
-                        $attributeValidationModel =
-                            $validationModel->attributes->properties->{$filterKey};
-                    } else {
-                        throw new \Exception(sprintf(
-                            'Attribute "%s" doesn\'t have a validation model',
-                            $filterKey
-                        ));
-                    }
-
-                    //$validationModelAttributes = $validationModel->attributes;
-
-                    $filterable = $modelClass::getFilterable();
-
-                    $isJSONFilter = false;
-
-                    //Check if $filterKeyParts and key contains . dot character
-                    if ($filterableJSON && strpos($filterKey, '.') !== false) {
-                        $filterKeyParts = explode('.', $filterKey);
-
-                        if (count($filterKeyParts) > 2) {
-                            throw new RequestException(
-                                'Second level filtering for JSON objects is not available'
-                            );
-                        }
-
-                        $filterSubkey = $filterKeyParts[1];
-
-                        //Hack check $filterSubkey if valid using regexp
-                        Validate::regexp(
-                            $filterSubkey,
-                            '/^[a-zA-Z_\-0-9]{1,30}$/',
-                            'filter[' . $filterKey . ']'
-                        );
-
-                        $filterKey = $filterKeyParts[0];
-
-                        $isJSONFilter = true;
-                    }
-
-                    if (!key_exists($filterKey, $filterable)) {
-                        throw new RequestException(sprintf(
-                            'Filter key "%s" not allowed',
-                            $filterKey
-                        ));
-                    }
-
-                    $operatorClass = $filterable[$filterKey];
-
-                    if ($isJSONFilter && ($operatorClass & Operator::CLASS_JSONOBJECT) === 0) {
-                        throw new RequestException(sprintf(
-                            'Filter key "%s" is not accepting JSON object filtering',
-                            $filterKey
-                        ));
-                    }
-
-                    //All must be arrays
-                    if (!is_array($filterValue)) {
-                        $filterValue = [$filterValue];
-                    }
-
-                    foreach ($filterValue as $singleFilterValue) {
-                        if (is_array($singleFilterValue)) {
-                            throw new RequestException(sprintf(
-                                'Array given for filter "%s"',
-                                $filterKey
-                            ));
-                        }
-
-                        $singleFilterValue = urldecode($singleFilterValue);
-
-                        list($operator, $operant) = Operator::parse($singleFilterValue);
-
-                        //Validate operator (check if it's in allowed operators class)
-                        if (!in_array(
-                            $operator,
-                            Operator::getByClassFlags($operatorClass)
-                        )) {
-                            throw new RequestException(sprintf(
-                                'Not allowed operator for field "%s"',
-                                $filterKey
-                            ));
-                        }
-
-                        if ((in_array($operator, Operator::getNullableOperators()))) {
-                            //Do nothing for nullable operators
-                        } else {
-                            //if (!$validationModelAttributes
-                            //    || !isset($validationModelAttributes->properties->{$filterKey})
-                            //) {
-                            //    throw new \Exception(sprintf(
-                            //        'Attribute "%s" doesn\'t have a validation model',
-                            //        $filterKey
-                            //    ));
-                            //}
-
-                            if ($isJSONFilter) {
-                                //unparsable
-                            } else {
-                                //use filterValidationModel for this property
-                                //if defined and operator is a CLASS_LIKE operator
-                                //if (in_array($operator, Operator::getLikeOperators())
-                                //    && $filterValidationModel
-                                //    && isset($filterValidationModel->properties->{$filterKey})
-                                //) {
-                                //    //Validate operant value
-                                //    $operant = $filterValidationModel->properties
-                                //        ->{$filterKey}->parse($operant);
-                                //} else {
-                                //    //Validate operant value
-                                //    $operant = $validationModelAttributes->properties
-                                //        ->{$filterKey}->parse($operant);
-                                //}
-                                //
-                                $operant = $attributeValidationModel->parse($operant);
-                            }
-                        }
-                        if ($isJSONFilter) {
-                            //Push tuple to attribute filters
-                            $filter->attributesJSON[] = [$filterKey, $filterSubkey, $operator, $operant];
-                        } else {
-                            //Push tuple to attribute filters
-                            $filter->attributes[] = [$filterKey, $operator, $operant];
-                        }
-                    }
-                }
-
-            }
+            $filter = Filter::parseFromParameters(
+                $params,
+                $modelClass
+            );
         }
 
-        //Parse pagination
-        if (isset($params->page)) {
-            $page = new Page();
-
-            if (isset($params->page['offset'])) {
-                $page->offset =
-                    (new \Phramework\Validate\UnsignedIntegerValidator())
-                        ->parse($params->page['offset']);
-            }
-
-            if (isset($params->page['limit'])) {
-                $page->limit =
-                    (new \Phramework\Validate\UnsignedIntegerValidator())
-                        ->parse($params->page['limit']);
-            }
-        }
-
-        //Push pagination $page object to end of arguments
-        $additionalGetArguments[] = $page;
-
-        if ($filterable) {
-            //Push filters to end of arguments
-            $additionalGetArguments[] = $filter;
-        }
+        $page = Page::parseFromParameters($params);
 
         if ($sortable) {
-            $modelSort = $modelClass::getSort();
-
-            $sort = null;
-
-            if ($modelSort->default !== null) {
-                $sort = new Sort();
-                $sort->table = $modelClass::getTable();
-                $sort->attribute = $modelSort->default;
-                $sort->ascending = $modelSort->ascending;
-
-                //Don't accept arrays
-                if (isset($params->sort)) {
-                    if (!is_string($params->sort)) {
-                        throw new RequestException(
-                            'String expected for sort'
-                        );
-                    }
-
-                    $validateExpression = sprintf(
-                        '/^(?P<descending>\-)?(?P<attribute>%s)$/',
-                        implode('|', $modelSort->attributes)
-                    );
-
-                    if (!!preg_match($validateExpression, $params->sort, $matches)) {
-                        $sort->attribute = $matches['attribute'];
-                        $sort->ascending = (
-                            isset($matches['descending']) && $matches['descending']
-                            ? false
-                            : true
-                        );
-
-                    } else {
-                        throw new RequestException(
-                            'Invalid value for sort'
-                        );
-                    }
-                }
-            }
-
-            //Push sort to end of arguments
-            $additionalGetArguments[] = $sort;
+            $sort = Sort::parseFromParameters(
+                $params,
+                $modelClass
+            );
         }
 
-        $data = call_user_func_array(
-            [$modelClass, 'get'],
-            $additionalGetArguments
+        $data = $modelClass::get(
+            $page,
+            $filter,
+            $sort,
+            null,//fields
+            $primaryDataParameters
         );
 
         $requestInclude = static::getRequestInclude($params);
@@ -350,7 +96,7 @@ abstract class GET extends \Phramework\JSONAPI\Controller\GETById
         $includedData = $modelClass::getIncludedData(
             $data,
             $requestInclude,
-            $additionalRelationshipsArguments
+            $relationshipsParameters
         );
 
         return static::viewData(
